requirements.txt
eth-ape==0.7.23
numpy==1.26.4
pony==0.7.17
psycopg2==2.9.9